requirements.txt
# -*- coding: utf-8 -*-
# Dependências principais do Flask e extensões
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Mail>=0.9.1
Flask-WTF>=1.0.0 # Para formulários seguros
email-validator>=1.1.3 # Validação de e-mail para WTForms

# Bibliotecas de Astrologia e Geocoding
flatlib>=0.2.0 # Biblioteca principal de astrologia
# geopy>=2.2.0 # Descomente se for implementar geocoding real via API

# Interação com API Externa
openai>=0.27.0 # Para integração com OpenAI (verifique a versão mais recente)
google-generativeai

# Geração de PDF
reportlab>=3.6.0

# Utilitários
python-dotenv>=0.19.0 # Para carregar variáveis de ambiente de .env

# Servidor WSGI (para produção)
gunicorn>=20.1.0 # Ou outro como waitress

